feat: 지하철 구간 등록시 구간의 길이가 노선의 길이보다 작거나 같을 경우 예외 처리 구현

+++ src/main/java/subway/controller/dto/SectionCreateRequest.java
@@ -12,12 +12,12 @@ public class SectionCreateRequest {
     private String downStationId;
     @NotBlank
     private String upStationId;
-    private int distance;
+    private long distance;
 
     public SectionCreateRequest() {
     }
 
-    public SectionCreateRequest(String downStationId, String upStationId, int distance) {
+    public SectionCreateRequest(String downStationId, String upStationId, long distance) {
         this.downStationId = downStationId;
         this.upStationId = upStationId;
         this.distance = distance;
@@ -35,7 +35,7 @@ public class SectionCreateRequest {
         return upStationId;
     }
 
-    public int getDistance() {
+    public long getDistance() {
         return distance;
     }
 }
